requirements.txt
numpy>=1.21.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
pytest>=7.0.0
httpx>=0.24.0
typing-extensions>=4.0.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
python-dotenv>=1.0.1
